refactor: add ParserBagiraService
